// client/src/js/bindings/ApiConfigCheckResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApiConfigResponse } from "./ApiConfigResponse";

export type ApiConfigCheckResponse = { config: ApiConfigResponse, client_version: string, compatible: boolean, client_too_old: boolean, };